Refactors mocks, adds promise handling to jest
